+++ backend/src/models/picker_options/index.js
@@ -290,6 +290,32 @@ const tableLookupValues = (projectId, contractId) => {
               FROM data.project_deliverable prjd
               WHERE project_id = ${projectId}
             ) projbudgdelname)`,
+      },
+      {
+        id: "programArea",
+        name: "program_area_option",
+        title: "Program Area",
+        description: "the project budget program areas",
+        table: "",
+        value: "",
+        label: "",
+        queryAdditions: ``,
+        customDefinition: `
+          (
+            SELECT COALESCE(json_agg(programArea), '[]')
+            FROM (
+              SELECT DISTINCT ON (financial_contact_name)
+                cc.program_area,
+                con.first_name || ' ' || con.last_name AS financial_contact_name,
+                min.ministry_short_name,
+                pb.client_coding_id AS value
+              FROM data.project_budget pb
+              LEFT JOIN data.client_coding cc ON pb.client_coding_id = cc.id
+              LEFT JOIN data.contact con ON cc.contact_id = con.id
+              LEFT JOIN data.ministry min ON con.ministry_id = min.id              
+              WHERE cc.project_id = ${projectId}
+            ) as programArea
+          )`,
       }
     );
   }