+++ CHANGELOG.md
@@ -1,10 +1,12 @@
+**December 14, 2023 [DESCW-1702](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1702)**
 
-
+- Frontend Backend
+  - Added program area picker for project budget view/edit forms.
 
 **December 14, 2023 [DESCW-395](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-395)**
 
-- Frontend 
-    - Removed Sass from the app
+- Frontend
+  - Removed Sass from the app
 
 **December 14, 2023 [DESCW-1259](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1259)**